# lsif/docs/api.yaml
openapi: 3.0.0
info:
  title: LSIF Server
  description: An internal Sourcegraph microservice that serves LSIF-powered code intelligence.
  version: 1.0.0
  contact:
    name: Eric Fritz
    email: eric@sourcegraph.com
    url: https://sourcegraph.com
servers:
  - url: http://localhost:3186
...
paths:
...
  /uploads/{repository}:
    get:
      description: Get LSIF uploads by state.
      tags:
        - Uploads
      parameters:
        - name: repository
          in: query
          description: The repository name.
          required: true
          schema:
            type: string
        - name: query
          in: query
          description: A search query applied over repository, commit, root, failure reason, and failure stacktrace properties.
          required: false
          schema:
            type: string
        - name: state
          in: query
          description: The target upload state.
          required: true
          schema:
            type: string
            enum:
              - processing
              - errored
              - completed
              - queued
        - name: visibleAtTip
          in: query
          description: If true, only show dumps visible at tip.
          required: false
          schema:
            type: boolean
        - name: limit
          in: query
          description: The maximum number of uploads to return in one page.
          required: false
          schema:
            type: number
            default: 50
        - name: offset
          in: query
          description: The number of uploads seen on previous pages.
          required: false
          schema:
            type: number
            default: 0
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Uploads'
          headers:
            Link:
              description: If there are more results, this header includes the URL of the next page with relation type *next*. See [RFC 5988](https://tools.ietf.org/html/rfc5988).
              schema:
                type: string
  /uploads/{id}:
    get:
      description: Get an LSIF upload by its identifier.
      tags:
        - Uploads
      parameters:
        - name: id
          in: path
          description: The upload identifier.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Upload'
        '404':
          description: Not Found
    delete:
      description: Delete an LSIF upload by its identifier.
      tags:
        - Uploads
      parameters:
        - name: id
          in: path
          description: The upload identifier.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: No Content
        '404':
          description: Not Found
components:
  schemas:
...
    Uploads:
      type: object
      description: A paginated wrapper for a list of uploads.
      properties:
        uploads:
          type: array
          description: A list of uploads with a particular state.
          items:
            $ref: '#/components/schemas/Upload'
        totalCount:
          type: number
          description: The total number of uploads in this set of results.
      required:
        - uploads
      additionalProperties: false
    Upload:
      type: object
      description: An LSIF upload.
      properties:
        id:
          type: string
          description: A unique identifier.
        repository:
          type: string
          description: The repository argument given on upload.
        commit:
          type: string
          description: The commit argument given on upload.
        root:
          type: string
          description: The root argument given on upload.
        filename:
          type: string
          description: The filename where the upload was stored before conversion.
        state:
          type: string
          description: The upload's current state.
          enum:
            - processing
            - errored
            - completed
            - queued
        failureSummary:
          type: string
          description: A brief description of why the upload conversion failed.
        failureStacktrace:
          type: string
          description: The stacktrace of the upload error.
        uploadedAt:
          type: string
          description: An ISO 8601-formatted time that the dump was uploaded.
        startedAt:
          type: string
          description: An ISO 8601-formatted time that the conversion started.
          nullable: true
        finishedAt:
          type: string
          description: An ISO 8601-formatted time that the conversion completed or errored.
          nullable: true
        visibleAtTip:
          type: boolean
          description: Whether or not this dump can provide global reference code intelligence.
      required:
        - id
        - repository
        - commit
        - root
        - filename
        - state
        - failureSummary
        - failureStacktrace
        - uploadedAt
        - startedAt
        - finishedAt
      additionalProperties: false